+++ src/main/java/datekeeper/datekeeperbackend/controllers/ApiController.java
@@ -1,12 +1,11 @@
-package datekeeper.datekeeperbackend.controller;
+package datekeeper.datekeeperbackend.controllers;
 
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 
 @RestController
-public class controller {
+public class ApiController {
     @GetMapping("/hello")
     public String HelloWorld() {
         return "Hello-World ";
