+++ src/test/java/datekeeper/datekeeperbackend/modells/EventTests.java
@@ -1,4 +1,4 @@
-package datekeeper.datekeeperbackend.domain;
+package datekeeper.datekeeperbackend.modells;
 
 public class EventTests {
     // TODO add test here
